+++ p2p_wallet/Scenes/Sell/Service/SellActionServiceProvider/SellActionServiceProvider.swift
@@ -1,0 +1,12 @@
+import Foundation
+
+public protocol SellActionServiceProvider {
+    associatedtype Quote: SellActionServiceQuote
+
+    func sellQuote(
+        baseCurrencyCode: String,
+        quoteCurrencyCode: String,
+        baseCurrencyAmount: Double,
+        extraFeePercentage: Double
+    ) async throws -> Quote
+}
